third_party/WebKit/LayoutTests/external/wpt/css/selectors/invalidation/matches.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Selectors Invalidation: :any-link</title>
    <link rel="author" title="Victoria Su" href="mailto:victoriaytsu@google.com">
    <link rel="help" href="https://drafts.csswg.org/selectors-4/#matches">
    <meta name="assert" content="This tests that the :matches() selector is effective">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
      .b {
        color: rgb(255, 255, 0); /* yellow */
      }
      /*Simple selector arguments */
      .a :matches(.b, .c) {
        color: rgb(255, 0, 0); /* red */
      }
      /*Compound selector arguments */
      .a :matches(.c#d, .e) {
        color: rgb(0, 255, 0); /* green */
      }
      /* Complex selector arguments */
      .a .g>.b {
        color: rgb(0, 0, 0); /* black */
      }
      .a :matches(.e+.f, .g>.b, .h) {
        color: rgb(0, 0, 255); /* blue */
      }
      .g>.b {
        color: rgb(0, 0, 0); /* black */
      }
      .a .h {
        color: rgb(0, 0, 0); /* black */
      }
      /* Nested */
      .a+.c>.e {
        color: rgb(0, 0, 0); /* black */
      }
      .a+:matches(.b+.f, :matches(.c>.e, .g)) {
        color: rgb(255, 0, 0); /* red */
      }
      .c>.e {
        color: rgb(0, 0, 0); /* black */
      }
    </style>
  </head>
  <body>
    <div id="a1">
      <div class="b" id="b1">
        Red
      </div>
      <div class="c" id="c1">
        Red
      </div>
      <div class="c" id="d">
        Green
      </div>
      <div class="e" id="e1">
        Green
      </div>
      <div class="f" id="f1">
        Blue
      </div>
      <div class="g">
        <div class="b" id="b2">
          Blue
          <div class="b" id="b3">
            Red
          </div>
        </div>
      </div>
      <div class="h" id="h1">
        Black
      </div>
    </div>
    <div class="c" id="c2">
      <div class="e" id="e2">
        Red
      </div>
    </div>
    <script>
      document.body.offsetTop;

      test(() => {
        assert_equals(getComputedStyle(b1).color, "rgb(255, 255, 0)");
        assert_equals(getComputedStyle(b2).color, "rgb(0, 0, 0)");
        assert_equals(getComputedStyle(b3).color, "rgb(255, 255, 0)");
        assert_equals(getComputedStyle(c1).color, "rgb(0, 0, 0)");
        assert_equals(getComputedStyle(d).color, "rgb(0, 0, 0)");
        assert_equals(getComputedStyle(e1).color, "rgb(0, 0, 0)");
        assert_equals(getComputedStyle(e2).color, "rgb(0, 0, 0)");
        assert_equals(getComputedStyle(f1).color, "rgb(0, 0, 0)");
        assert_equals(getComputedStyle(h1).color, "rgb(0, 0, 0)");
      }, "Preconditions.");

      test(() => {
        a1.className = "a";
        assert_equals(getComputedStyle(b1).color, "rgb(255, 0, 0)");
        assert_equals(getComputedStyle(b3).color, "rgb(255, 0, 0)");
        assert_equals(getComputedStyle(c1).color, "rgb(255, 0, 0)");
      }, "Invalidate :matches() for simple selector arguments.");

      test(() => {
        a1.className = "a";
        assert_equals(getComputedStyle(d).color, "rgb(0, 255, 0)");
      }, "Invalidate :matches() for compound selector arguments.");

      test(() => {
        a1.className = "a";
        assert_equals(getComputedStyle(b2).color, "rgb(0, 0, 255)");
        assert_equals(getComputedStyle(b3).color, "rgb(255, 0, 0)");
        assert_equals(getComputedStyle(f1).color, "rgb(0, 0, 255)");
      }, "Invalidate :matches() for complex selector arguments.");

      test(() => {
        a1.className = "a";
        assert_equals(getComputedStyle(e2).color, "rgb(255, 0, 0)");
      }, "Invalidate nested :matches().");

      test(() => {
        a1.className = "a";
        assert_equals(getComputedStyle(b2).color, "rgb(0, 0, 255)");
        assert_equals(getComputedStyle(h1).color, "rgb(0, 0, 0)");
      }, "Test specificity of :matches().");
    </script>
  </body>
</html>
